+++ contracts/transfer/tests/common/utils.rs
@@ -6,22 +6,33 @@
 
 use std::sync::mpsc;
 
+use rusk_abi::{
+    ContractError, ContractId, EconomicMode, Error, Session, TRANSFER_CONTRACT,
+};
+
 use dusk_plonk::prelude::*;
 use poseidon_merkle::Opening as PoseidonOpening;
 
 use execution_core::{
     transfer::{TreeLeaf, TRANSFER_TREE_DEPTH},
-    Note, Transaction, ViewKey,
-};
-use rusk_abi::{
-    CallReceipt, ContractError, ContractId, EconomicMode, Error, Session,
-    TRANSFER_CONTRACT,
+    BlsScalar, Note, Transaction, ViewKey,
 };
 
-const POINT_LIMIT: u64 = 0x100000000;
+const POINT_LIMIT: u64 = 0x10_000_000;
 
-const H: usize = TRANSFER_TREE_DEPTH;
-const A: usize = 4;
+pub struct ExecutionResult {
+    pub gas_spent: u64,
+    pub economic_mode: EconomicMode,
+}
+
+impl ExecutionResult {
+    pub const fn new(gas_spent: u64, economic_mode: EconomicMode) -> Self {
+        Self {
+            gas_spent,
+            economic_mode,
+        }
+    }
+}
 
 pub fn leaves_from_height(
     session: &mut Session,
@@ -33,7 +44,7 @@ pub fn leaves_from_height(
         TRANSFER_CONTRACT,
         "leaves_from_height",
         &height,
-        u64::MAX,
+        POINT_LIMIT,
         feeder,
     )?;
 
@@ -41,6 +52,32 @@ pub fn leaves_from_height(
         .iter()
         .map(|bytes| rkyv::from_bytes(&bytes).expect("Should return leaves"))
         .collect())
+}
+
+pub fn leaves_from_pos(
+    session: &mut Session,
+    pos: u64,
+) -> Result<Vec<TreeLeaf>, Error> {
+    let (feeder, receiver) = mpsc::channel();
+
+    session.feeder_call::<_, ()>(
+        TRANSFER_CONTRACT,
+        "leaves_from_pos",
+        &pos,
+        POINT_LIMIT,
+        feeder,
+    )?;
+
+    Ok(receiver
+        .iter()
+        .map(|bytes| rkyv::from_bytes(&bytes).expect("Should return leaves"))
+        .collect())
+}
+
+pub fn num_notes(session: &mut Session) -> Result<u64, Error> {
+    session
+        .call(TRANSFER_CONTRACT, "num_notes", &(), u64::MAX)
+        .map(|r| r.data)
 }
 
 pub fn update_root(session: &mut Session) -> Result<(), Error> {
@@ -55,10 +92,19 @@ pub fn root(session: &mut Session) -> Result<BlsScalar, Error> {
         .map(|r| r.data)
 }
 
+pub fn module_balance(
+    session: &mut Session,
+    contract: ContractId,
+) -> Result<u64, Error> {
+    session
+        .call(TRANSFER_CONTRACT, "module_balance", &contract, POINT_LIMIT)
+        .map(|r| r.data)
+}
+
 pub fn opening(
     session: &mut Session,
     pos: u64,
-) -> Result<Option<PoseidonOpening<(), H, A>>, Error> {
+) -> Result<Option<PoseidonOpening<(), TRANSFER_TREE_DEPTH, 4>>, Error> {
     session
         .call(TRANSFER_CONTRACT, "opening", &pos, POINT_LIMIT)
         .map(|r| r.data)
@@ -80,25 +126,17 @@ pub fn prover_verifier(circuit_name: &str) -> (Prover, Verifier) {
     (prover, verifier)
 }
 
-pub fn filter_notes_owned_by<I: IntoIterator<Item = Note>>(
-    vk: ViewKey,
-    iter: I,
-) -> Vec<Note> {
-    iter.into_iter().filter(|note| vk.owns(note)).collect()
-}
-
-/// Executes a transaction, returning the call receipt
+/// Executes a transaction.
+/// Returns result containing gas spent and economic mode.
 pub fn execute(
     session: &mut Session,
     tx: Transaction,
-) -> Result<CallReceipt<Result<Vec<u8>, ContractError>>, Error> {
-    // Spend the inputs and execute the call. If this errors the transaction is
-    // unspendable.
+) -> Result<ExecutionResult, Error> {
     let mut receipt = session.call::<_, Result<Vec<u8>, ContractError>>(
         TRANSFER_CONTRACT,
         "spend_and_execute",
         &tx,
-        tx.fee.gas_limit,
+        u64::MAX,
     )?;
 
     // Ensure all gas is consumed if there's an error in the contract call
@@ -106,9 +144,11 @@ pub fn execute(
         receipt.gas_spent = receipt.gas_limit;
     }
 
-    // Refund the appropriate amount to the transaction. This call is guaranteed
-    // to never error. If it does, then a programming error has occurred. As
-    // such, the call to `Result::expect` is warranted.
+    let contract_id = tx
+        .call
+        .clone()
+        .map(|(module_id, _, _)| ContractId::from_bytes(module_id));
+
     let refund_receipt = session
         .call::<_, ()>(
             TRANSFER_CONTRACT,
@@ -116,8 +156,8 @@ pub fn execute(
             &(
                 tx.fee,
                 receipt.gas_spent,
-                EconomicMode::None,
-                None::<ContractId>,
+                receipt.economic_mode.clone(),
+                contract_id,
             ),
             u64::MAX,
         )
@@ -125,5 +165,16 @@ pub fn execute(
 
     receipt.events.extend(refund_receipt.events);
 
-    Ok(receipt)
+    Ok(ExecutionResult::new(
+        receipt.gas_spent,
+        receipt.economic_mode,
+    ))
+}
+
+/// Returns vector of notes owned by a given view key.
+pub fn filter_notes_owned_by<I: IntoIterator<Item = Note>>(
+    vk: ViewKey,
+    iter: I,
+) -> Vec<Note> {
+    iter.into_iter().filter(|note| vk.owns(note)).collect()
 }