+++ node/src/database/rocksdb.rs
@@ -877,6 +877,7 @@ mod tests {
 
     use fake::{Fake, Faker};
     use node_data::ledger::Transaction;
+    use rusk_abi::EconomicMode;
 
     #[test]
     fn test_store_block() {
@@ -1104,6 +1105,7 @@ mod tests {
                 inner: t.clone(),
                 block_height: 0,
                 gas_spent: 0,
+                economic_mode: EconomicMode::None,
                 err: None,
             })
             .collect()